requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
discord.py==2.3.2
distro==1.8.0
dnspython==2.4.2
docopt==0.6.2
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
iso8601==2.1.0
multidict==6.0.4
openai==1.1.1
pipreqs==0.4.13
pydantic==2.4.2
pydantic_core==2.13.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rfc3986==2.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tqdm==4.66.1
twitchAPI==4.0.1
typing_extensions==4.8.0
urllib3==2.0.7
yarg==0.1.9
yarl==1.9.2
